Move initialize function for mag disturbance
